examples/RANO/data_preparator/project/requirements.txt
typer==0.9.0
pandas==1.5.3
PyYAML==6.0.1
# Include all your requirements here
SimpleITK==2.3.1
tqdm==4.66.2
scikit-image==0.21.0
FigureGenerator==0.0.4
gandlf==0.0.16
labelfusion==1.0.14
nibabel==5.1.0